backend/requirements.txt
livekit-agents[cartesia,silero,openai]
python-dotenv
fastapi
uvicorn
livekit-server-sdk
python-multipart